refactor(python): apply Ludo-feedback cleanup (python/Justfile, examples, README) + root justfile delegation
